Removed ihkexport plugin Added ihkexportold plugin Added ihkexportnew plugin
Reason: Two different versions needed. Everything with "old" in name will be deleted after Jan-19
