--- plugins/org.projectforge.plugins.ihkexport/src/main/java/org/projectforge/plugins/ihkexport/IhkExportPFPluginService.java
@@ -1,81 +1,0 @@
-/////////////////////////////////////////////////////////////////////////////
-//
-// Project ProjectForge Community Edition
-//         www.projectforge.org
-//
-// Copyright (C) 2001-2019 Micromata GmbH, Germany (www.micromata.com)
-//
-// ProjectForge is dual-licensed.
-//
-// This community edition is free software; you can redistribute it and/or
-// modify it under the terms of the GNU General Public License as published
-// by the Free Software Foundation; version 3 of the License.
-//
-// This community edition is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General
-// Public License for more details.
-//
-// You should have received a copy of the GNU General Public License along
-// with this program; if not, see http://www.gnu.org/licenses/.
-//
-/////////////////////////////////////////////////////////////////////////////
-
-package org.projectforge.plugins.ihkexport;
-
-import org.projectforge.plugins.core.AbstractPlugin;
-import org.projectforge.plugins.core.PFPluginService;
-
-/**
- * The Class MemoPFPluginService.
- *
- * @author Roger Rene Kommer (r.kommer.extern@micromata.de)
- */
-public class IhkExportPFPluginService implements PFPluginService
-{
-
-  /**
-   * {@inheritDoc}
-   *
-   */
-
-  @Override
-  public String getPluginId()
-  {
-    return "ihkexport";
-  }
-
-  /**
-   * {@inheritDoc}
-   *
-   */
-
-  @Override
-  public String getPluginName()
-  {
-    return "IHK-Export";
-  }
-
-  /**
-   * {@inheritDoc}
-   *
-   */
-
-  @Override
-  public String getPluginDescription()
-  {
-    return "Ausbildungsnachweise drucken";
-  }
-
-  /**
-   * {@inheritDoc}
-   *
-   */
-
-  @Override
-  public AbstractPlugin createPluginInstance()
-  {
-    return new IhkExportPlugin();
-  }
-
-}
