--- plugins/org.projectforge.plugins.ihkexport/src/main/resources/IhkExportI18nResources.properties
@@ -1,5 +1,0 @@
-plugins.ihkexport.title=IHKExport
-plugins.ihkexport.menu=IHKExport
-plugins.ihkexport.download=Download
-plugins.ihkexport.noitemsfound=No items found for export
-plugins.ihkexport.title.heading=IHKExport
