--- plugins/org.projectforge.plugins.ihkexport/src/main/resources/IhkExportI18nResources_de.properties
@@ -1,5 +1,0 @@
-plugins.ihkexport.title=IHKExport
-plugins.ihkexport.menu=IHKExport
-plugins.ihkexport.download=Herunterladen
-plugins.ihkexport.noitemsfound=Keine Elemente zum Exportieren gefunden
-plugins.ihkexport.title.heading=IHKExport
