--- plugins/org.projectforge.plugins.ihkexport/src/main/resources/META-INF/services/org.projectforge.framework.persistence.api.RightRightIdProviderService
@@ -1,2 +1,0 @@
-org.projectforge.plugins.ihkexport.IhkExportPluginUserRightId$ProviderService
-
