--- plugins/org.projectforge.plugins.ihkexport/src/main/resources/META-INF/services/org.projectforge.plugins.core.PFPluginService
@@ -1,1 +1,0 @@
-org.projectforge.plugins.ihkexport.IhkExportPFPluginService
