+++ plugins/org.projectforge.plugins.ihkexportnew/src/main/resources/IhkExportNewI18nResources.properties
@@ -1,0 +1,5 @@
+plugins.ihkexportNew.title=ihkexportNewN
+plugins.ihkexportNew.menu=ihkexportNew
+plugins.ihkexportNew.download=Download
+plugins.ihkexportNew.noitemsfound=No items found for export
+plugins.ihkexportNew.title.heading=ihkexportNew
