+++ plugins/org.projectforge.plugins.ihkexportnew/src/main/resources/IhkExportNewI18nResources_de.properties
@@ -1,0 +1,5 @@
+plugins.ihkexportNew.title=IHKExportNeu
+plugins.ihkexportNew.menu=IHKExportNeu
+plugins.ihkexportNew.download=Herunterladen
+plugins.ihkexportNew.noitemsfound=Keine Elemente zum Exportieren gefunden
+plugins.ihkexportNew.title.heading=IHKExportNeu
