+++ plugins/org.projectforge.plugins.ihkexportnew/src/main/resources/META-INF/services/org.projectforge.framework.persistence.api.RightRightIdProviderService
@@ -1,0 +1,2 @@
+org.projectforge.plugins.ihkexportNew.IhkExportPluginUserRightIdNew$ProviderService
+
