+++ plugins/org.projectforge.plugins.ihkexportnew/src/main/resources/META-INF/services/org.projectforge.plugins.core.ProjectforgePluginService
@@ -1,0 +1,1 @@
+org.projectforge.plugins.ihkexportNew.IhkExportProjectforgePluginServiceNew
