+++ plugins/org.projectforge.plugins.ihkexportold/src/main/resources/IhkExportOldI18nResources.properties
@@ -1,0 +1,5 @@
+plugins.ihkexportOld.title=ihkexportOld
+plugins.ihkexportOld.menu=ihkexportOld
+plugins.ihkexportOld.download=Download
+plugins.ihkexportOld.noitemsfound=No items found for export
+plugins.ihkexportOld.title.heading=ihkexportOld
