+++ plugins/org.projectforge.plugins.ihkexportold/src/main/resources/IhkExportOldI18nResources_de.properties
@@ -1,0 +1,5 @@
+plugins.ihkexportOld.title=ihkexportAlt
+plugins.ihkexportOld.menu=ihkexportAlt
+plugins.ihkexportOld.download=Herunterladen
+plugins.ihkexportOld.noitemsfound=Keine Elemente zum Exportieren gefunden
+plugins.ihkexportOld.title.heading=ihkexportAlt
