+++ plugins/org.projectforge.plugins.ihkexportold/src/main/resources/IhkExportOldI18nResources_en.properties
@@ -1,0 +1,1 @@
+# Empty, en is default.
