+++ plugins/org.projectforge.plugins.ihkexportold/src/main/resources/META-INF/services/org.projectforge.framework.persistence.api.RightRightIdProviderService
@@ -1,0 +1,2 @@
+org.projectforge.plugins.ihkexportOld.IhkExportPluginUserRightIdOld$ProviderService
+
