+++ projectforge-application/pom.xml
@@ -63,7 +63,13 @@
     </dependency>
     <dependency>
       <groupId>org.projectforge</groupId>
-      <artifactId>org.projectforge.plugins.ihkexport</artifactId>
+      <artifactId>org.projectforge.plugins.ihkexportold</artifactId>
+      <version>${project.version}</version>
+      <scope>runtime</scope>
+    </dependency>
+    <dependency>
+      <groupId>org.projectforge</groupId>
+      <artifactId>org.projectforge.plugins.ihkexportnew</artifactId>
       <version>${project.version}</version>
       <scope>runtime</scope>
     </dependency>
